# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RV Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/saved:
    get:
      tags: [user]
      operationId: getSavedListings
      summary: Get user saved listings
      responses:
        '200':
          description: Saved listings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedListingsResponse'
        '401':
          description: Unauthorized
  /listings/{id}/save:
    post:
      tags: [user]
      operationId: saveListing
      summary: Save a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Listing saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    delete:
      tags: [user]
      operationId: unsaveListing
      summary: Unsave a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Listing unsaved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /user/searches:
    get:
      tags: [user]
      operationId: getSavedSearches
      summary: Get saved searches
      responses:
        '200':
          description: Saved searches
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedSearchesResponse'
    post:
      tags: [user]
      operationId: saveSearch
      summary: Save a search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveSearchRequest'
      responses:
        '200':
          description: Search saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /user/searches/{id}:
    delete:
      tags: [user]
      operationId: deleteSavedSearch
      summary: Delete a saved search
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Search deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /user/alerts:
    get:
      tags: [user]
      operationId: getPriceAlerts
      summary: Get price alerts
      responses:
        '200':
          description: Price alerts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceAlertsResponse'
    post:
      tags: [user]
      operationId: createPriceAlert
      summary: Create a price alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePriceAlertRequest'
      responses:
        '200':
          description: Alert created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /user/alerts/{id}:
    delete:
      tags: [user]
      operationId: deletePriceAlert
      summary: Delete a price alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Alert deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /user/messages:
    get:
      tags: [user]
      operationId: getMessages
      summary: Get sent messages
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
    post:
      tags: [user]
      operationId: sendMessage
      summary: Send a message to a dealer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  schemas:
...
    DealScore:
      type: string
      enum:
        - great_deal
        - good_deal
        - fair_deal
        - high_price
        - overpriced
    RvType:
      type: string
      enum:
        - class_a
        - class_b
        - class_c
        - travel_trailer
        - fifth_wheel
        - toy_hauler
        - popup_camper
        - truck_camper
    Listing:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: number
        type:
          $ref: "#/components/schemas/RvType"
        price:
          type: number
        marketValue:
          type: number
        dealScore:
          $ref: "#/components/schemas/DealScore"
        dealSavings:
          type: number
        mileage:
          type: number
        length:
          type: number
        slides:
          type: number
        location:
          type: string
        state:
          type: string
        dealerName:
          type: string
        dealerId:
          type: number
        images:
          type: array
          items:
            type: string
        daysOnMarket:
          type: number
        sleeps:
          type: number
        condition:
          type: string
        isNew:
          type: boolean
        isFeatured:
          type: boolean
        vin:
          type: string
        pricePerFoot:
          type: number
      required:
        - id
        - title
        - make
        - model
        - year
        - type
        - price
        - marketValue
        - dealScore
        - location
        - state
        - dealerName
        - dealerId
        - images
        - daysOnMarket
        - sleeps
        - condition
        - isNew
...
    SuccessResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    SavedListingsResponse:
      type: object
      required: [listings]
      properties:
        listings:
          type: array
          items:
            $ref: '#/components/schemas/Listing'
...
    SaveSearchRequest:
      type: object
      required: [name, filters]
      properties:
        name:
          type: string
        filters:
          type: object
          additionalProperties: true
    SavedSearchItem:
      type: object
      required: [id, name, filters, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        filters:
          type: object
          additionalProperties: true
        createdAt:
          type: string
    SavedSearchesResponse:
      type: object
      required: [searches]
      properties:
        searches:
          type: array
          items:
            $ref: '#/components/schemas/SavedSearchItem'
    CreatePriceAlertRequest:
      type: object
      required: [listingId, targetPrice]
      properties:
        listingId:
          type: number
        targetPrice:
          type: number
    PriceAlertItem:
      type: object
      required: [id, listingId, targetPrice, triggered, createdAt]
      properties:
        id:
          type: number
        listingId:
          type: number
        targetPrice:
          type: number
        triggered:
          type: boolean
        createdAt:
          type: string
        listingTitle:
          type: string
        currentPrice:
          type: number
    PriceAlertsResponse:
      type: object
      required: [alerts]
      properties:
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/PriceAlertItem'
    SendMessageRequest:
      type: object
      required: [dealerId, listingId, body]
      properties:
        dealerId:
          type: number
        listingId:
          type: number
        body:
          type: string
    MessageItem:
      type: object
      required: [id, dealerId, listingId, body, read, createdAt]
      properties:
        id:
          type: number
        dealerId:
          type: number
        dealerName:
          type: string
        listingId:
          type: number
        listingTitle:
          type: string
        body:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
    MessagesResponse:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/MessageItem'
...